Rejoin orphan letters to their parent words
## Problem

Many words contained a gap, occurring just after the first letter.

## Solution

Add a new processor that searches through the document
for single letters.

If the following word is not valid,
and the combination of the orphan letter with the word is valid,
then combine them.

There are also small checks that take into account:

* the article `a`, as in `a choice`
* single letters used as ordinal categories,
  such as in `subchapters B and C are ...`
* the possessive `s`,
  where the apostrophe can be confused as whitespace.

## Additional Changes

Add a small skeleton for a debugging layer.

There were a couple of elusive bugs which were hard to track down –
the document-level diffs didn't provide enough information.

It was  really difficult to see why things weren't being processed correctly.