Regroup characters assigned to wrong word
## Problem

There are some instances,
with two words next to each other,
where the first letter of the second word
comes across as tacked onto the end of the first word.

## Solution

Regroup the letters into their correct words.

This is a pretty expensive operation,
and requires the processor to spell-check each word in the document.

## Side effects

This processor does not respect specific gaps between words –
whenever it finds whitespace between two words,
it combines it all into a single space.

This is desirable in the final output of the document,
but it makes the document diffs for this step especially noisy.